requirements.txt
about-time==4.2.1
alive-progress==3.1.4
async-timeout==4.0.3
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
chosic==0.0.5
cloudscraper==1.2.71
customtkinter==5.2.0
darkdetect==0.8.0
fake-useragent==1.2.1
filelock==3.12.4
grapheme==0.6.0
idna==3.4
kcu==0.0.73
ksimpleapi==0.0.41
noraise==0.0.16
pyparsing==3.1.1
python-dotenv==1.0.0
ratelimit==2.2.1
redis==5.0.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
simple-multiprocessing==0.0.15
six==1.16.0
soupsieve==2.5
spotipy==2.23.0
stopit==1.1.2
tldextract==3.6.0
urllib3==2.0.4
